Move `AddProductFromImageTextFieldView`'s view model to a separate class `AddProductFromImageTextFieldViewModel`.
